Rename, Move : 클래스명과 파일 위치 변경
- 클래스 이름을 명사로 수정했다.
- common 폴더의 하위 폴더에 하나의 파일만 있는 경우, common 폴더로 옮겼다.
