+++ src/main/java/common/InputValidator.java
@@ -1,4 +1,4 @@
-package common.validate;
+package common;
 
 import common.exception.EmptyFormException;
 
@@ -6,7 +6,7 @@ import static common.util.Util.split;
 import static webserver.RequestParser.PARAMETERS_DELIMITER;
 import static webserver.RequestParser.PARAMETER_DELIMITER;
 
-public class InputValidate {
+public class InputValidator {
 
     public static void validateUserInfo(String query) {
         String[] paramters = split(query, PARAMETERS_DELIMITER);
