+++ src/main/java/dto/HttpRequest.java
@@ -4,7 +4,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Map;
 
-import static common.view.OutputView.printRequest;
+import static common.Logger.printRequest;
 import static webserver.RequestParser.*;
 
 public class HttpRequest {
