+++ src/test/java/common/binder/BinderTest.java
@@ -1,5 +1,6 @@
 package common.binder;
 
+import common.Binder;
 import model.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
